fix keyboard not opening
